add feature flag and move old table
