+++ packages/web/components/complex/index.ts
@@ -1,3 +1,2 @@
-export * from "./all-pools-table";
 export * from "./paginated-table";
 export * from "./pool";
